Instructions and Goals
The README.md contains the script for this workshop in a way that's
hopefully useful for someone to follow along with at their own pace.

In addition, this commit includes a possible solution for two of the
exercieses laid out in the README.

It does NOT include the template directory or any of the details
mentioned there, so that the git history can make it clear what parts of
that template are stock and which parts were changed as part of the
development of the course.